move `candidate_from_obligation` out of assembly
it doesn't belong there as it also does winnowing
